Refactoring how memory management works to simplify the experience and allow suballocation

+++ sources/Graphics/Core/Graphics/GraphicsAdapter.cs
@@ -33,7 +33,7 @@ public abstract unsafe class GraphicsAdapter : GraphicsServiceObject
     /// <summary>Creates a new graphics device which utilizes the adapter.</summary>
     /// <param name="createMemoryAllocator">A function pointer to a method which creates the backing memory allocators used by the device or <c>null</c> to use the system provided default memory allocator.</param>
     /// <exception cref="ObjectDisposedException">The adapter has been disposed.</exception>
-    public abstract GraphicsDevice CreateDevice(delegate*<GraphicsDeviceObject, ulong, GraphicsMemoryAllocator> createMemoryAllocator);
+    public abstract GraphicsDevice CreateDevice(delegate*<GraphicsDeviceObject, nuint, GraphicsMemoryAllocator> createMemoryAllocator);
 
     /// <inheritdoc />
     public override string ToString() => Name;
