+++ sources/Graphics/D3D12/Graphics/D3D12GraphicsAdapter.cs
@@ -76,7 +76,7 @@ public sealed unsafe class D3D12GraphicsAdapter : GraphicsAdapter
     public override uint VendorId => DxgiAdapterDesc.VendorId;
 
     /// <inheritdoc />
-    public override D3D12GraphicsDevice CreateDevice(delegate*<GraphicsDeviceObject, ulong, GraphicsMemoryAllocator> createMemoryAllocator)
+    public override D3D12GraphicsDevice CreateDevice(delegate*<GraphicsDeviceObject, nuint, GraphicsMemoryAllocator> createMemoryAllocator)
     {
         ThrowIfDisposedOrDisposing(_state, nameof(D3D12GraphicsAdapter));
         return new D3D12GraphicsDevice(this, createMemoryAllocator);
